+++ src/BigInteger.php
@@ -1026,13 +1026,13 @@ final readonly class BigInteger extends BigNumber
     /**
      * Returns the index of the rightmost (lowest-order) one bit in this BigInteger.
      *
-     * Returns -1 if this BigInteger contains no one bits.
+     * Returns null if this BigInteger is zero.
      *
-     * @return -1|non-negative-int
+     * @return non-negative-int|null
      *
      * @pure
      */
-    public function getLowestSetBit(): int
+    public function getLowestSetBit(): ?int
     {
         $n = $this;
         $bitLength = $this->getBitLength();
@@ -1045,7 +1045,7 @@ final readonly class BigInteger extends BigNumber
             $n = $n->shiftedRight(1);
         }
 
-        return -1;
+        return null;
     }
 
     /**